modified script to compare L_DDPM between source and target domains, fixed bug in backpropr through x_cond for L_DDPM

+++ tta_uia_segmentation/src/models/ConditionalGaussianDiffusion.py
@@ -70,7 +70,7 @@ class ConditionalGaussianDiffusion(GaussianDiffusion):
         max_t = default(max_t, self.num_timesteps)
         t = torch.randint(min_t, max_t, (b,), device=device).long()
         img = self.normalize(img)
-        cond_img = self.normalize(cond_img.type(torch.int8))  
+        cond_img = self.normalize(cond_img)  
         
         if img.max() > 1 or img.min() < -1:
             print('Warning: img is not normalized between -1 and 1'
